refactor: fetchPostsByTag 로직 분리
