+++ src/entities/post/actions/fetchPosts.ts
