+++ src/entities/post/api/post.ts
@@ -10,6 +10,16 @@ export const getPosts = async (limit: number, skip: number): Promise<PostRespons
   }
 }
 
+export const getPostsByTag = async (tag: string): Promise<PostResponse> => {
+  try {
+    const res = await fetch(`/api/posts/tag/${tag}`)
+    return res.json()
+  } catch (e) {
+    console.error("태그별 게시물 가져오기 오류:", e)
+    return { posts: [], total: 0 }
+  }
+}
+
 export const fetchSearchPosts = async (searchQuery: string): Promise<PostResponse> => {
   try {
     const res = await fetch(`/api/posts/search?q=${searchQuery}`)
